move attention_decoder function to decoder
