+++ neuralmonkey/decoding_function.py
@@ -8,67 +8,7 @@ See http://arxiv.org/abs/1606.07481
 
 import tensorflow as tf
 
-from neuralmonkey.logging import debug
 from neuralmonkey.nn.projection import linear
-
-# pylint: disable=too-many-arguments,too-many-locals
-# Great functions require great number of parameters
-def attention_decoder(decoder_inputs, initial_state, attention_objects,
-                      cell, loop_function=None, scope=None,
-                      summary_collections=None):
-    outputs = []
-    states = []
-
-    #### WTF does this do?
-    # do manualy broadcasting of the initial state if we want it
-    # to be the same for all inputs
-    if len(initial_state.get_shape()) == 1:
-        debug("Warning! I am doing this weird job.")
-        state_size = initial_state.get_shape()[0].value
-        initial_state = tf.reshape(tf.tile(initial_state,
-                                           tf.shape(decoder_inputs[0])[:1]),
-                                   [-1, state_size])
-
-    with tf.variable_scope(scope or "attention_decoder"):
-        output, state = decode_step_nomaxout(decoder_inputs[0], initial_state,
-                                    attention_objects, cell)
-        outputs.append(output)
-        states.append(state)
-
-        for step in range(1, len(decoder_inputs)):
-            tf.get_variable_scope().reuse_variables()
-
-            if loop_function:
-                decoder_input = loop_function(output, step)
-            else:
-                decoder_input = decoder_inputs[step]
-
-            output, state = decode_step_nomaxout(decoder_input, state,
-                                                 attention_objects, cell)
-            outputs.append(output)
-            states.append(state)
-
-        if summary_collections:
-            for i, a in enumerate(attention_objects):
-                attentions = a.attentions_in_time[-len(decoder_inputs):]
-                alignments = tf.expand_dims(tf.transpose(
-                    tf.pack(attentions), perm=[1, 2, 0]), -1)
-
-                tf.image_summary("attention_{}".format(i), alignments,
-                                 collections=summary_collections,
-                                 max_images=256)
-
-    return outputs, states
-
-
-def decode_step_nomaxout(prev_output, prev_state, attention_objects,
-                         rnn_cell):
-    contexts = [a.attention(prev_state) for a in attention_objects]
-    output = tf.concat(1, [prev_output, prev_state] + contexts)
-    _, state = rnn_cell(tf.concat(1, [prev_output] + contexts), prev_state)
-
-    return output, state
-
 
 # def decode_step(prev_output, prev_state, attention_objects,
 #                 rnn_cell, maxout_size):
